Limit the prefetch size of rows and columns of sheet.
Sometimes we can try to process a file with a very large numbers of rows
and/or columns, which can make our program to get stuck scanning the
file.

To avoid this, it could be interesting to offer the possibility to,
previously to the scan of the file in the AsDataSet method, inform via a
parameter in the configuration object of a max numbers of columns and
rows to read and project only the inside results into the ouput dataset.

This is even more interesting in case where you are interested, for
example, only in the headers and the first rows, though the sheet can
contain hundreds or thousands of values, but not relevant to the scan.